fix: move in memory exporter behind testing feature flag
