+++ opentelemetry-sdk/src/testing/metrics/mod.rs
@@ -1,0 +1,3 @@
+pub use in_memory_exporter::{InMemoryMetricsExporter, InMemoryMetricsExporterBuilder};
+
+mod in_memory_exporter;
