+++ opentelemetry-sdk/src/testing/mod.rs
@@ -1,2 +1,5 @@
 #[cfg(all(feature = "testing", feature = "trace"))]
 pub mod trace;
+
+#[cfg(all(feature = "testing", feature = "metrics"))]
+pub mod metrics;
